third_party/WebKit/LayoutTests/platform/mac/fast/dom/Range/getClientRects-expected.txt

Test 1
PASS rects.length is 5
PASS rects[0].left is 8
PASS rects[0].top is 8
PASS rects[0].width is 400
PASS rects[0].height is 160
PASS rects[1].left is 8
PASS rects[1].top is 19
FAIL rects[1].width should be 396. Was 386.578125.
PASS rects[1].height is 18
PASS rects[2].left is 8
PASS rects[2].top is 59
FAIL rects[2].width should be 398. Was 359.90625.
PASS rects[2].height is 18
PASS rects[3].left is 8
PASS rects[3].top is 99
FAIL rects[3].width should be 360. Was 381.25.
PASS rects[3].height is 18
PASS rects[4].left is 8
PASS rects[4].top is 139
FAIL rects[4].width should be 306. Was 368.359375.
PASS rects[4].height is 18
Test 2
PASS rects.length is 4
PASS rects[0].left is 8
PASS rects[0].top is 215
FAIL rects[0].width should be 396. Was 386.578125.
PASS rects[0].height is 18
PASS rects[1].left is 8
PASS rects[1].top is 255
FAIL rects[1].width should be 398. Was 359.90625.
PASS rects[1].height is 18
PASS rects[2].left is 8
PASS rects[2].top is 295
FAIL rects[2].width should be 360. Was 381.25.
PASS rects[2].height is 18
PASS rects[3].left is 8
PASS rects[3].top is 335
FAIL rects[3].width should be 306. Was 368.359375.
PASS rects[3].height is 18
Test 3
PASS rects.length is 4
FAIL rects[0].left should be 80. Was 81.
PASS rects[0].top is 411
FAIL rects[0].width should be 324. Was 313.578125.
PASS rects[0].height is 18
PASS rects[1].left is 8
PASS rects[1].top is 451
FAIL rects[1].width should be 398. Was 359.90625.
PASS rects[1].height is 18
PASS rects[2].left is 8
PASS rects[2].top is 491
FAIL rects[2].width should be 360. Was 381.25.
PASS rects[2].height is 18
PASS rects[3].left is 8
PASS rects[3].top is 531
FAIL rects[3].width should be 88. Was 147.
PASS rects[3].height is 18
Test 4
PASS rects.length is 7
PASS rects[0].left is 8
PASS rects[0].top is 596
PASS rects[0].width is 400
PASS rects[0].height is 278
PASS rects[1].left is 8
PASS rects[1].top is 607
FAIL rects[1].width should be 396. Was 386.578125.
PASS rects[1].height is 18
PASS rects[2].left is 8
PASS rects[2].top is 725
FAIL rects[2].width should be 242. Was 268.390625.
PASS rects[2].height is 18
FAIL rects[3].left should be 326. Was 352.390625.
PASS rects[3].top is 725
FAIL rects[3].width should be 44. Was 44.875.
PASS rects[3].height is 18
PASS rects[4].left is 8
PASS rects[4].top is 765
FAIL rects[4].width should be 341. Was 348.375.
PASS rects[4].height is 18
PASS rects[5].left is 8
PASS rects[5].top is 805
FAIL rects[5].width should be 366. Was 375.890625.
PASS rects[5].height is 18
PASS rects[6].left is 8
PASS rects[6].top is 845
FAIL rects[6].width should be 67. Was 67.96875.
PASS rects[6].height is 18
Test 5
PASS rects.length is 5
FAIL rects[0].left should be 80. Was 81.
PASS rects[0].top is 921
FAIL rects[0].width should be 324. Was 313.578125.
PASS rects[0].height is 18
PASS rects[1].left is 8
PASS rects[1].top is 1039
FAIL rects[1].width should be 242. Was 268.390625.
PASS rects[1].height is 18
FAIL rects[2].left should be 250. Was 276.390625.
PASS rects[2].top is 950
PASS rects[2].width is 76
PASS rects[2].height is 103
FAIL rects[3].left should be 326. Was 352.390625.
PASS rects[3].top is 1039
FAIL rects[3].width should be 44. Was 44.875.
PASS rects[3].height is 18
PASS rects[4].left is 8
PASS rects[4].top is 1079
FAIL rects[4].width should be 284. Was 290.
PASS rects[4].height is 18
Test 6
FAIL rects.length should be 7. Was 8.
PASS rects[0].left is 8
PASS rects[0].top is 1224
PASS rects[0].width is 400
FAIL rects[0].height should be 200. Was 240.
PASS rects[1].left is 8
PASS rects[1].top is 1235
FAIL rects[1].width should be 396. Was 386.578125.
PASS rects[1].height is 18
PASS rects[2].left is 8
PASS rects[2].top is 1275
FAIL rects[2].width should be 57. Was 16.
PASS rects[2].height is 18
FAIL rects[3].left should be 105. Was 8.
FAIL rects[3].top should be 1275. Was 1315.
FAIL rects[3].width should be 44. Was 58.21875.
PASS rects[3].height is 18
FAIL rects[4].left should be 8. Was 106.21875.
PASS rects[4].top is 1315
FAIL rects[4].width should be 387. Was 45.3125.
PASS rects[4].height is 18
PASS rects[5].left is 8
PASS rects[5].top is 1355
FAIL rects[5].width should be 397. Was 396.34375.
PASS rects[5].height is 18
PASS rects[6].left is 8
PASS rects[6].top is 1395
FAIL rects[6].width should be 171. Was 359.921875.
PASS rects[6].height is 18
Test 7
FAIL rects.length should be 7. Was 8.
PASS rects[0].left is 8
FAIL rects[0].top should be 1471. Was 1511.
FAIL rects[0].width should be 396. Was 386.578125.
PASS rects[0].height is 18
PASS rects[1].left is 8
FAIL rects[1].top should be 1500. Was 1551.
FAIL rects[1].width should be 400. Was 16.
FAIL rects[1].height should be 40. Was 18.
PASS rects[2].left is 8
FAIL rects[2].top should be 1511. Was 1580.
FAIL rects[2].width should be 57. Was 400.
FAIL rects[2].height should be 18. Was 40.
FAIL rects[3].left should be 105. Was 8.
FAIL rects[3].top should be 1511. Was 1591.
FAIL rects[3].width should be 44. Was 58.21875.
PASS rects[3].height is 18
FAIL rects[4].left should be 8. Was 106.21875.
FAIL rects[4].top should be 1551. Was 1591.
FAIL rects[4].width should be 387. Was 45.3125.
PASS rects[4].height is 18
PASS rects[5].left is 8
FAIL rects[5].top should be 1591. Was 1631.
FAIL rects[5].width should be 397. Was 396.34375.
PASS rects[5].height is 18
PASS rects[6].left is 8
FAIL rects[6].top should be 1631. Was 1671.
FAIL rects[6].width should be 171. Was 359.921875.
PASS rects[6].height is 18
Test 8
PASS rects.length is 4
FAIL rects[0].left should be 168. Was 168.5107421875.
FAIL rects[0].top should be 1678. Was 1758.438720703125.
FAIL rects[0].width should be 269. Was 262.27642822265625.
FAIL rects[0].height should be 316. Was 307.706298828125.
FAIL rects[1].left should be 137. Was 137.8689727783203.
FAIL rects[1].top should be 1704. Was 1784.1502685546875.
FAIL rects[1].width should be 271. Was 245.13206481933594.
FAIL rects[1].height should be 317. Was 287.2742919921875.
FAIL rects[2].left should be 107. Was 107.22718811035156.
FAIL rects[2].top should be 1729. Was 1809.86181640625.
FAIL rects[2].width should be 246. Was 258.8515625.
FAIL rects[2].height should be 289. Was 303.62451171875.
FAIL rects[3].left should be 76. Was 76.58541107177734.
FAIL rects[3].top should be 1755. Was 1835.5732421875.
FAIL rects[3].width should be 212. Was 250.56564331054688.
FAIL rects[3].height should be 247. Was 293.749755859375.
Test 9
FAIL rects.length should be 1. Was 0.
FAIL rects[0].left should be 8. Threw exception TypeError: Cannot read property 'left' of undefined
FAIL rects[0].top should be 1903. Threw exception TypeError: Cannot read property 'top' of undefined
FAIL rects[0].width should be 0. Threw exception TypeError: Cannot read property 'width' of undefined
FAIL rects[0].height should be 18. Threw exception TypeError: Cannot read property 'height' of undefined
Test 9b
PASS rects.length is 1
PASS rects[0].left is 8
FAIL rects[0].top should be 1903. Was 1983.
PASS rects[0].width is 0
PASS rects[0].height is 18
Test 10
FAIL rects.length should be 4. Was 5.
PASS rects[0].left is 19
FAIL rects[0].top should be 2088. Was 2168.
PASS rects[0].width is 18
FAIL rects[0].height should be 393. Was 375.
PASS rects[1].left is 59
FAIL rects[1].top should be 2088. Was 2168.
PASS rects[1].width is 18
FAIL rects[1].height should be 377. Was 367.03125.
PASS rects[2].left is 99
FAIL rects[2].top should be 2088. Was 2168.
PASS rects[2].width is 18
FAIL rects[2].height should be 372. Was 348.375.
PASS rects[3].left is 139
FAIL rects[3].top should be 2088. Was 2168.
PASS rects[3].width is 18
FAIL rects[3].height should be 360. Was 375.890625.
Test 11
PASS rects.length is 0
PASS successfullyParsed is true

TEST COMPLETE
